+++ backend/common/src/models/tracker/peer.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, Local};
 pub struct Peer {
     pub id: i64,
     pub user_id: i32,
-    pub torrent_id: i64,
+    pub torrent_id: i32,
     pub peer_id: [u8; 20],
     pub ip: Option<std::net::Ipv4Addr>,
     pub port: u16,
