+++ backend/storage/src/models/entity.rs
@@ -41,7 +41,7 @@ pub struct EntityLite {
 #[derive(Debug, Deserialize, Serialize, FromRow, ToSchema)]
 pub struct AffiliatedEntity {
     pub id: i64,
-    pub title_group_id: i64,
+    pub title_group_id: i32,
     pub entity_id: i64,
     pub created_by_id: i32,
     #[schema(value_type = String, format = DateTime)]
@@ -51,7 +51,7 @@ pub struct AffiliatedEntity {
 #[derive(Debug, Deserialize, Serialize, ToSchema)]
 pub struct AffiliatedEntityHierarchy {
     pub id: i64,
-    pub title_group_id: i64,
+    pub title_group_id: i32,
     pub entity_id: i64,
     pub created_by_id: i32,
     #[schema(value_type = String, format = DateTime)]
